requirements.txt
python-dotenv==1.0.1
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=4.3.0,<5.0.0
djoser>=2.1.0
drf-yasg==1.21.7
